requirements.txt
# GUI ----------------------------------------
chardet
qtawesome>=1.2.3
pyside6>=6.4.2
pyqtdarktheme>=2.1.0

# Base ----------------------------------------
matplotlib>=3.2.2
opencv-python>=4.6.0
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
pytorch-lightning==2.0.3
tqdm>=4.64.0
ultralytics
timm
xmltodict

# Visualization ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Extras --------------------------------------
psutil  # system utilization
thop>=0.1.1  # FLOPs computation